refactor(canvas): isolate migration recovery
